requirements.txt
beautifulsoup4==4.12.2
black==23.11.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
googlesearch-python==1.2.3
greenlet==3.0.0
idna==3.4
mypy-extensions==1.0.0
numpy==1.26.2
packaging==23.2
pandas==2.1.3
pathspec==0.11.2
platformdirs==4.0.0
playwright==1.39.0
playwright-stealth==1.0.6
pyee==11.0.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
soupsieve==2.5
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
